add package.d, implement Driver/Connection/ConnectionPool factory methods with improved URL support

+++ examples/pgsql_test/source/main.d
@@ -1,39 +1,17 @@
 import std.stdio;
 
-import ddbc.all;
-import ddbc.pods;
+import ddbc;
 
 int main(string[] argv)
 {
-    string[string] params;
-    // This part depends on RDBMS
-    version( USE_PGSQL )
-    {
-        PGSQLDriver driver = new PGSQLDriver();
-        //string url = PGSQLDriver.generateUrl( "/tmp", 5432, "ddbctestdb" ) ~ "?user=ddbctest,password=ddbctestpass";
-        string url = PGSQLDriver.generateUrl( "localhost", 5432, "ddbctestdb" ) ~ "?user=ddbctest,password=ddbctestpass";
-        //params["user"] = "ddbctest";
-        //params["password"] = "ddbctestpass";
-        params["ssl"] = "true";
-    }
-    else version( USE_SQLITE )
-    {
-        SQLITEDriver driver = new SQLITEDriver();
-        string url = "zzz.db"; // file with DB
-    }
-    else version(USE_MYSQL)
-    {
-        // MySQL driver - you can use PostgreSQL or SQLite instead as well
-        MySQLDriver driver = new MySQLDriver();
-        string url = MySQLDriver.generateUrl("localhost", 3306, "test_db");
-        params = MySQLDriver.setUserAndPassword("testuser", "testpassword");
-    }
-
-    // create connection pool
-    DataSource ds = new ConnectionPoolDataSourceImpl(driver, url, params);
+    string url = "ddbc:postgresql://localhost:5432/ddbctestdb?user=ddbctest,password=ddbctestpass,ssl=true";
+    //string url = "ddbc:mysql://lx4-vlopatin.nn.five9.com:3306/ddbctestdb?user=db59,password=admin";
+    //immutable string url = "ddbc:sqlite:testdb.sqlite";
+    immutable string driverName = extractDriverNameFromURL(url);
 
     // creating Connection
-    auto conn = ds.getConnection();
+    //auto conn = ds.getConnection();
+    Connection conn = createConnection(url);
     scope(exit) conn.close();
 
     // creating Statement
@@ -61,37 +39,43 @@ int main(string[] argv)
     while (rs.next())
         writeln(to!string(rs.getLong(1)), "\t", rs.getString(2), "\t", rs.getString(3), "\t", rs.getString(4));
 
-    ubyte[] bin_data = [1, 2, 3, 'a', 'b', 'c', 0xFE, 0xFF, 0, 1, 2];
-    stmt.executeUpdate("DROP TABLE IF EXISTS bintest");
-    stmt.executeUpdate("CREATE TABLE bintest (id bigint not null primary key, blob1 bytea)");
-    PreparedStatement ps = conn.prepareStatement("INSERT INTO bintest (id, blob1) VALUES (1, ?)");
-    ps.setUbytes(1, bin_data);
-    ps.executeUpdate();
-    struct Bintest {
-        long id;
-        ubyte[] blob1;
-    }
-    Bintest[] rows;
-    foreach(e; stmt.select!Bintest)
-        rows ~= e;
-    //stmt!
-    auto rs2 = stmt.executeQuery("SELECT id, blob1 FROM bintest WHERE id=1");
-    if (rs2.next()) {
-        ubyte[] res = rs2.getUbytes(2);
-        assert(res == bin_data);
+    // PostgreSQL bytea blobs test
+    if (driverName == "postgresql") {
+        ubyte[] bin_data = [1, 2, 3, 'a', 'b', 'c', 0xFE, 0xFF, 0, 1, 2];
+        stmt.executeUpdate("DROP TABLE IF EXISTS bintest");
+        stmt.executeUpdate("CREATE TABLE bintest (id bigint not null primary key, blob1 bytea)");
+        PreparedStatement ps = conn.prepareStatement("INSERT INTO bintest (id, blob1) VALUES (1, ?)");
+        ps.setUbytes(1, bin_data);
+        ps.executeUpdate();
+        struct Bintest {
+            long id;
+            ubyte[] blob1;
+        }
+        Bintest[] rows;
+        foreach(e; stmt.select!Bintest)
+            rows ~= e;
+        //stmt!
+        auto rs2 = stmt.executeQuery("SELECT id, blob1 FROM bintest WHERE id=1");
+        if (rs2.next()) {
+            ubyte[] res = rs2.getUbytes(2);
+            assert(res == bin_data);
+        }
     }
 
-    stmt.executeUpdate("DROP TABLE IF EXISTS guidtest");
-    stmt.executeUpdate("CREATE TABLE guidtest (guid uuid not null primary key, name text)");
-    stmt.executeUpdate("INSERT INTO guidtest (guid, name) VALUES ('cd3c7ffd-7919-f6c5-999d-5586d9f3b261', 'vasia')");
-    struct Guidtest {
-        string guid;
-        string name;
+    // PostgreSQL uuid type test
+    if (driverName == "postgresql") {
+        stmt.executeUpdate("DROP TABLE IF EXISTS guidtest");
+        stmt.executeUpdate("CREATE TABLE guidtest (guid uuid not null primary key, name text)");
+        stmt.executeUpdate("INSERT INTO guidtest (guid, name) VALUES ('cd3c7ffd-7919-f6c5-999d-5586d9f3b261', 'vasia')");
+        struct Guidtest {
+            string guid;
+            string name;
+        }
+        Guidtest[] guidrows;
+        foreach(e; stmt.select!Guidtest)
+            guidrows ~= e;
+        writeln(guidrows);
     }
-    Guidtest[] guidrows;
-    foreach(e; stmt.select!Guidtest)
-        guidrows ~= e;
-    writeln(guidrows);
 
     return 0;
 }
